Forgot to saved the numberOfRecords changes in the previous commit

+++ Excerciese/Excercise_2/TimedTask.h
@@ -9,7 +9,7 @@ namespace seneca
     constexpr size_t Max = 10;
     class TimedTask
     {
-        int numberOfRecords{};
+        size_t numberOfRecords{};
         std::chrono::time_point<std::chrono::steady_clock> start {};
         std::chrono::time_point<std::chrono::steady_clock> end{};
         struct task
